Playtesting fixes
Color trails and clip points ignoring the distance,
as distance can be seen by the position of the point itself
Lower the raycount in crossline calc by 4 times (from 180 to 45)
to improve performance
Fix the zone ratio reporting. In the future this may
need some refactoring.
Fix scoring window resizing values
Comment out debug drawing, a switch for it should be made in the future.